+++ rstapi/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # Copyright 2024 RST Cloud Pty Ltd
 """Python library for accessing RST Cloud API."""
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 __author__ = "RST Cloud Pty Ltd"
 
 from .connectivity import connectivity  # noqa
@@ -10,4 +10,5 @@ from .noisecontrol import noisecontrol  # noqa
 from .reporthub import reporthub  # noqa
 from .scan import scan  # noqa
 from .threatfeed import threatfeed  # noqa
+from .threatobjects import threatobjects  # noqa
 from .whoisapi import whoisapi  # noqa